Progress with saving

+++ src/main/java/com/example/navalbattle/controller/LoginController.java
@@ -1,32 +1,43 @@
 package com.example.navalbattle.controller;
 
+import com.example.navalbattle.model.PlainTextFileHandler;
 import com.example.navalbattle.view.FleetStage;
-import com.example.navalbattle.view.GameStage;
 import com.example.navalbattle.view.LoginStage;
-import com.example.navalbattle.view.WelcomeStage;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.scene.control.TextField;
 
-import java.io.IOException;
-
-
-public class LoginController  {
-
-
+public class LoginController {
 
     @FXML
     private TextField userTxt;
-
+    String userInput;
     @FXML
     void buttonPlayGame(ActionEvent event) {
-        LoginStage.deleteInstance();
-        FleetStage.getInstance();
+        if (userTxt != null && !userTxt.getText().isEmpty()) {
+            String userInput = userTxt.getText();
+            System.out.println("El usuario ingresó: " + userInput);
+
+            // Guardar el nombre de usuario en un archivo (opcional)
+            PlainTextFileHandler fileHandler = new PlainTextFileHandler();
+            fileHandler.writeToFile("usuario.txt", userInput);
+
+            // Configurar el nombre de usuario en GameController
+            GameController.setStaticUserName(userInput);
+
+            // Proceder con la eliminación de la instancia de LoginStage y abrir GameStage
+            LoginStage.deleteInstance();
+            FleetStage.getInstance();
+        } else {
+            System.out.println("No se ingresó ningún nombre de usuario.");
+        }
+
+
     }
 
     @FXML
     void handleClickExit(ActionEvent event) {
         LoginStage.deleteInstance();
-       // WelcomeStage.getInstance();
+        // WelcomeStage.getInstance(); // Opcional si decides reactivar esta línea
     }
 }
